fix bug on fedora by zeroing struct tm
Apparently strptime fills in the struct tm differently depending on the
state in which it's passed, such as returning times near the epoch (like a
-1 time), so make sure to set it to zero before calling strptime.  Thank
you valgrind, now the tests pass again on Fedora 32.

Be more deliberate about handling errors when parsing time units, since the
wrong basetime upsets all the times.  Make fillAircraftTrack() private
since the error checking is not obvious and it is not called outside the
class yet anyway.



git-svn-id: http://svn.eol.ucar.edu/svn/raf/trunk/kml_tools@9358 640d5228-2204-0410-8d71-8f46fa6850b4