Add a config value for time gap between flights
Current value is hardcoded at 12 hours.  acTrack2kml keeps all the data
internal and does not do a full refresh to the database...unless the
timestamp of the last point is at least this many hours old.

The pro's of this are that if the database gets whiped mid-flight, the
histrory is still maintained in acTrack2kml.

This value needed to be configurable as small UAV's can do multiple
flights in a day.  So 12 hours was two long.