port track rendering to recent selenium
the page rendering requires serving the html and javascript files through a
local temporary web server, but it works with both the firefox and chrome web
driver implementations, chrome is the default.  also modify the runtime
environment so selenium can find selenium-manager.  convert NOTES.txt to
markdown and update the kml rendering information.